Fix ARM+ARC EXC_BAD_ACCESS in glossy gradient drawing
Within drawGlossyGradient(), the shineStarColor and shineEndColor
CGColorRefs are bad pointers by the time they're passed into
drawLinearGradient(). This is because the UIColor from which they were
built has been released via ARC's “avoid the autorelease pool”
optimization. In testing, this optimization only seems to manifest
itself on iOS devices and not in the simulator.

For a thorough explanation of the issue, see the Big Nerd Ranch's
blog post, 'ARC Gotcha -- Unexpectedly Short Lifetimes'[1].

[1]: http://weblog.bignerdranch.com/296-arc-gotcha-unexpectedly-short-lifetimes/